backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-genai
google-cloud-firestore
google-cloud-storage
PyPDF2
python-docx
pydantic
websockets
